delete old biscalc
