+++ backend/osrs-dmg-sim/biscalc/gearoptions.go
@@ -1,137 +1,178 @@
 package biscalc
 
-import "github.com/Maurits825/osrs-damage-sim-go/backend/osrs-damage-sim/dpscalc"
+import (
+	"errors"
+	"strings"
 
-const (
-	slayerHelm = 25912
-
-	infernal  = 21295
-	torture   = 19553
-	avernic   = 22322
-	ferocious = 22981
-	prims     = 13239
-
-	scythe     = 22325
-	fang       = 26219
-	rapier     = 22324
-	saladBlade = 23995
-
-	bandosGodsword  = 11804
-	armadylGodsword = 11802
-	abbysalDagger   = 13265
-	dragonDagger    = 1215
-	crystalHalberd  = 23987
-	voidwaker       = 27690
-	dragonClaws     = 13652
-	dragonWarhammer = 13576
-
-	ultor    = 28307
-	bellator = 28316
-
-	torvaHelm = 26382
-	torvaBody = 26384
-	torvaLegs = 26386
-
-	inqHelm = 24419
-	inqBody = 24420
-	inqLegs = 24421
+	"github.com/Maurits825/osrs-damage-sim-go/backend/osrs-damage-sim/dpscalc"
 )
 
-const (
-	anguish        = 19547
-	blessedQuiver  = 28955
-	zaryteVambs    = 26235
-	venatorRing    = 28310
-	pegs           = 13237
-	twistedBuckler = 21000
-
-	tbow           = 20997
-	blowpipe       = 12926
-	dragonDarts    = 11230
-	zaryteCrossbow = 26374
-	bowfa          = 25865
-
-	masoriHelm = 27235
-	masoriBody = 27238
-	masoriLegs = 27241
-
-	crystalHelm = 23971
-	crystalBody = 23975
-	crystalLegs = 23979
-
-	dragonArrows    = 11212
-	rubyDragonBolts = 21944
-)
-
-const (
-	occult        = 12002
-	mageArenaCape = 21795
-	tormented     = 19544
-	magusRing     = 28313
-	eternal       = 13235
-	elidinisWard  = 27251
-
-	shadowStaff = 27275
-	sangStaff   = 22323
-	harmStaff   = 24423
-	tomeOfFire  = 20714
-
-	ancestralHat  = 21018
-	ancestralBody = 21021
-	ancestralLegs = 21024
-
-	virtusHat  = 26241
-	virtusBody = 26243
-	virtusLegs = 26245
-)
-
-//TODO otherGear sets like void, always better to wear the set + other otherGear
-//TODO also salve if undead? or always... but kinda of a waste when we get more and more items
-
-type gearOptions map[dpscalc.GearSlot][]int
-type gearSetupOption map[dpscalc.GearSlot]dpscalc.GearItem //TODO gearOption
-
-var meleeGearOptions = gearOptions{
-	dpscalc.Head:   {torvaHelm, inqHelm},
-	dpscalc.Cape:   {infernal},
-	dpscalc.Neck:   {torture},
-	dpscalc.Body:   {torvaBody, inqBody},
-	dpscalc.Shield: {avernic},
-	dpscalc.Legs:   {torvaLegs, inqLegs},
-	dpscalc.Hands:  {ferocious},
-	dpscalc.Feet:   {prims},
-	dpscalc.Ring:   {ultor, bellator},
+var allGearSlots = []dpscalc.GearSlot{
+	dpscalc.Head,
+	dpscalc.Cape,
+	dpscalc.Neck,
+	dpscalc.Weapon,
+	dpscalc.Body,
+	dpscalc.Shield,
+	dpscalc.Legs,
+	dpscalc.Hands,
+	dpscalc.Feet,
+	dpscalc.Ring,
+	dpscalc.Ammo,
 }
 
-var meleeWeapons = []int{scythe, fang, rapier, saladBlade}
-var meleeSpecWeapons = []int{fang, bandosGodsword, armadylGodsword, abbysalDagger, crystalHalberd, voidwaker, dragonClaws, dragonWarhammer}
+type gearSetupOptions map[dpscalc.GearSlot][]int
+type gearSetup map[dpscalc.GearSlot]dpscalc.GearItem
 
-var rangedGearOptions = gearOptions{
-	dpscalc.Head:   {masoriHelm, crystalHelm},
-	dpscalc.Cape:   {blessedQuiver},
-	dpscalc.Neck:   {anguish},
-	dpscalc.Body:   {masoriBody, crystalBody},
-	dpscalc.Shield: {twistedBuckler},
-	dpscalc.Legs:   {masoriLegs, crystalLegs},
-	dpscalc.Hands:  {zaryteVambs},
-	dpscalc.Feet:   {pegs},
-	dpscalc.Ring:   {venatorRing},
-	dpscalc.Ammo:   {dragonArrows, rubyDragonBolts},
+func (gear gearSetup) isValid() bool {
+	weaponId := gear[dpscalc.Weapon].Id
+	isAmmoEmpty := gear[dpscalc.Ammo].Id == dpscalc.EmptyItemId
+
+	// TODO add more charged based ranged weapons
+	if weaponId == blowpipe || weaponId == bowfa {
+		if !isAmmoEmpty {
+			return false
+		}
+	}
+
+	//check 2h and shield
+	is2h := allItems[weaponId].Is2h
+	if is2h && gear[dpscalc.Shield].Id != dpscalc.EmptyItemId {
+		return false
+	}
+
+	//invalidate one handed weapon with no shield
+	//TODO for budget setups could allow this
+	if !is2h && gear[dpscalc.Shield].Id == dpscalc.EmptyItemId {
+		return false
+	}
+
+	//check ammo, TODO a bit scuffed
+	ammoName := allItems[gear[dpscalc.Ammo].Id].Name
+	switch allItems[weaponId].WeaponCategory {
+	case "CROSSBOW":
+		if !strings.Contains(ammoName, "bolts") {
+			return false
+		}
+	case "BOW":
+		if !strings.Contains(ammoName, "arrow") {
+			return false
+		}
+	case "THROWN", "CHINCHOMPAS":
+		if !isAmmoEmpty {
+			return false
+		}
+	}
+
+	//TODO other checks?
+	return true
 }
 
-var rangedWeapons = []int{tbow, blowpipe, zaryteCrossbow, bowfa}
-var rangedSpecWeapons = []int{blowpipe, zaryteCrossbow}
-
-var magicGearOptions = gearOptions{
-	dpscalc.Head:   {ancestralHat, virtusHat},
-	dpscalc.Cape:   {mageArenaCape},
-	dpscalc.Neck:   {occult},
-	dpscalc.Body:   {ancestralBody, virtusBody},
-	dpscalc.Shield: {elidinisWard, tomeOfFire},
-	dpscalc.Legs:   {ancestralLegs, virtusLegs},
-	dpscalc.Hands:  {tormented},
-	dpscalc.Feet:   {eternal},
-	dpscalc.Ring:   {magusRing},
+func (setup gearSetup) clone() gearSetup {
+	newSetup := make(gearSetup)
+	for _, slot := range allGearSlots {
+		newSetup[slot] = setup[slot]
+	}
+	return newSetup
 }
 
-var magicWeapons = []int{shadowStaff, sangStaff, harmStaff}
+func (options gearSetupOptions) enrichGearSetupOptions(style AttackStyle, setup *BisCalcInputSetup) {
+	options.addGearId(dpscalc.Shield, dpscalc.EmptyItemId)
+	options.addGearId(dpscalc.Ammo, dpscalc.EmptyItemId)
+
+	if style == Ranged {
+		options.addGearIds(dpscalc.Ammo, rangedAmmo)
+	}
+
+	//add weapons based on style
+	if setup.IsSpecialAttack {
+		options.addGearIds(dpscalc.Weapon, specWeapons[style])
+	} else {
+		options.addGearIds(dpscalc.Weapon, weapons[style])
+	}
+
+	if setup.IsOnSlayerTask {
+		options.addGearId(dpscalc.Head, slayerHelm)
+	}
+
+	npc := dpscalc.GetNpc(setup.GlobalSettings.Npc.Id)
+	if npc.IsUndead {
+		options.addGearId(dpscalc.Neck, salveAmuletEI)
+	}
+
+	//TODO could also add dragon bane weapons if dragon
+	//demon bane if demon and stuff
+}
+
+func (opt gearSetupOptions) addGearId(slot dpscalc.GearSlot, id int) {
+	opt[slot] = append(opt[slot], id)
+}
+
+func (opt gearSetupOptions) addGearIds(slot dpscalc.GearSlot, ids []int) {
+	opt[slot] = append(opt[slot], ids...)
+}
+
+// TODO simple func to get bis without budget consideration for now, from the bis graph
+func getGearSetupOptions(graph SlotBisGraph) gearSetupOptions {
+	options := make(gearSetupOptions)
+	for _, slot := range allGearSlots {
+		options[slot] = make([]int, len(graph[slot]))
+		for i := range graph[slot] {
+			options[slot][i] = graph[slot][i].ids[0] //take first item here
+		}
+	}
+	return options
+}
+
+func gearSetupOptionsIterator(options gearSetupOptions) func() (gearSetup, error) {
+	var currentGear gearSetup = make(gearSetup)
+	iteratorEmpty := false
+
+	var gearIndices map[dpscalc.GearSlot]int = make(map[dpscalc.GearSlot]int)
+	var gearLengths map[dpscalc.GearSlot]int = make(map[dpscalc.GearSlot]int)
+	for _, slot := range allGearSlots {
+		ids, ok := options[slot]
+		gearIndices[slot] = 0
+		gearLengths[slot] = len(ids)
+		if !ok {
+			gearLengths[slot] = 1
+		}
+	}
+
+	gearSlotIndex := len(allGearSlots) - 1
+	lastGearSlotIndex := gearSlotIndex
+
+	return func() (gearSetup, error) {
+		if iteratorEmpty {
+			return nil, errors.New("iterator empty")
+		}
+
+		//get current gear based on indices
+		for _, slot := range allGearSlots {
+			ids, ok := options[slot]
+			if ok {
+				id := ids[gearIndices[slot]]
+				currentGear[slot] = dpscalc.GearItem{Id: id}
+			}
+		}
+
+		//update indices
+		gearIndices[allGearSlots[lastGearSlotIndex]] += 1
+		gearSlotIndex = lastGearSlotIndex
+		gearSlot := allGearSlots[gearSlotIndex]
+		for gearIndices[gearSlot] == gearLengths[gearSlot] {
+			//iterator is empty
+			if gearSlotIndex == 0 {
+				iteratorEmpty = true
+				return currentGear, nil
+			}
+
+			gearIndices[gearSlot] = 0
+			gearSlotIndex--
+			gearSlot = allGearSlots[gearSlotIndex]
+			gearIndices[gearSlot] += 1
+		}
+
+		return currentGear, nil
+	}
+}
